splitting up into more components
