+++ src/Breakpoints.js
@@ -1,6 +1,11 @@
 export const media = {
-  mobile: "(mix-width: 768px)",
-  tablet: "(mix-width: 1024px)",
-  smalldesktop: "(mix-width: 1020px)",
-  desktop: "(mix-width: 1600px)",
+  mobile: "(min-width: 768px)",
+  tablet: "(min-width: 1024px)",
+  smalldesktop: "(min-width: 1280px)",
+  desktop: "(min-width: 1536px)",
 };
+
+//mobile = 768px = md
+//tablet = 1024px = lg
+//smalldesktop = 1280px = xl
+//desktop = 1536px = 2xl
