+++ src/sections/FeaturedProjects.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import { Title } from "./Title";
+import { Title } from "./TitleComponents/Title";
 import { ProjectBox } from "./FeaturedProjectsComponents/ProjectBox";
-import { SeeMore } from "./SeeMore";
+import { SeeMore } from "./SeeMoreComponent/SeeMore";
 
 export const FeaturedProjects = () => {
   return (
