+++ src/sections/SeeMoreComponent/SeeMore.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import seeMore from "../assets/Ic-ArrowDown.svg";
+import seeMore from "../../assets/Ic-ArrowDown.svg";
 
 const StyledSeeMore = styled.button`
   padding: 0px 16px;
